+++ src/main/java/damon/backend/exception/NotFountException.java
@@ -1,0 +1,8 @@
+package damon.backend.exception;
+
+public class NotFountException extends RuntimeException{
+    
+        public NotFountException(String message) {
+            super(message);
+        }
+}
